# Some speed gain for `inferences()`